move chat_completion to vllm_inference.py
