+++ modal_server_config.py
@@ -5,6 +5,7 @@ from pydantic import field_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 gpu_config_mapping = {
+    "meetkai/functionary-small-v2.5": modal.gpu.L4(count=1),
     "meetkai/functionary-small-v2.4": modal.gpu.L4(count=1),
     "meetkai/functionary-medium-v2.4": modal.gpu.A100(count=2, size="80GB"),
     "meetkai/functionary-small-v2.2": modal.gpu.L4(count=1),
@@ -16,7 +17,7 @@ gpu_config_mapping = {
 
 
 class Settings(BaseSettings):
-    model: str = "meetkai/functionary-small-v2.4"
+    model: str = "meetkai/functionary-small-v2.5"
     max_model_length: int = 8192
     gpu_config: Any = None  # Define the type according to what modal.gpu.* returns
     gpu_memory_utilization: float = 0.9
